+++ src/main/java/insane96mcp/iguanatweaksreborn/setup/ITClientConfig.java
@@ -2,7 +2,6 @@ package insane96mcp.iguanatweaksreborn.setup;
 
 import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.ClientModules;
-import insane96mcp.iguanatweaksreborn.module.Modules;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -11,7 +10,7 @@ import org.apache.commons.lang3.tuple.Pair;
 
 @Mod.EventBusSubscriber(modid = IguanaTweaksReborn.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class ITClientConfig {
-	public static final ForgeConfigSpec CLIENT_SPEC;
+	public static final ForgeConfigSpec CONFIG_SPEC;
 	public static final ClientConfig CLIENT;
 
 	public static final ForgeConfigSpec.Builder builder;
@@ -20,7 +19,7 @@ public class ITClientConfig {
 		builder = new ForgeConfigSpec.Builder();
 		final Pair<ClientConfig, ForgeConfigSpec> specPair = builder.configure(ClientConfig::new);
 		CLIENT = specPair.getLeft();
-		CLIENT_SPEC = specPair.getRight();
+		CONFIG_SPEC = specPair.getRight();
 	}
 
 	public static class ClientConfig {